agent/requirements-test.txt
imapclient
httpx
pyyaml
beautifulsoup4
python-dotenv
pytest
